Squashed commit of the following:
commit 33347b6315967d13b6e697191ea36c76a9717659
Author: Andrei Lazaro <alazaro.apps@gmail.com>
Date:   Fri Aug 18 03:49:43 2023 +0800

    Update register.blade.php

commit 76ecdfa87b9e86791bdaf9d9364a42f3736f1bc1
Author: Andrei Lazaro <alazaro.apps@gmail.com>
Date:   Fri Aug 18 03:38:35 2023 +0800

    Added Password reset, controlled registration, upgraded aesthetics, refined UI/UX, and dynamic password rules.

    - [NEW] Forgot Password Option: Users can now reset their passwords with the self-service password reset feature.
    - [NEW] User Registration with Approval: A user registration feature that requires Administrator approval, ensuring controlled access.
    - [UPDATE] Dynamic Password Format Requirements: Administrators now have the ability to set password rules, such as minimum character count, uppercase requirements, special character necessity, and more.
    - [UPDATE] Fresh Look and Functionality: The system has been upgraded to Metronic v8.2.0, offering improved functionality and a sleek new appearance.
    - [UPDATE] Enhanced Authentication UI/UX: Smoother login experience with the refined UI/UX design.

commit 9dcd50b022cede3b59ab941ef4338d141f46e12e
Author: Andrei Lazaro <alazaro.apps@gmail.com>
Date:   Wed Aug 16 05:14:31 2023 +0800

    Update version display from v0.3.0-alpha to v0.6.0-alpha

commit 30cf110dd0527775dea80f1c875e08ad4f14f104
Author: Andrei Lazaro <alazaro.apps@gmail.com>
Date:   Wed Aug 16 04:59:40 2023 +0800

    Profile enhancements, archive indicator.

    - [NEW] Archive Status Indicator: Quickly identify whether a student's profile is archived on the student profile view.
    - [NEW] Enhanced Student Profile View: Addition of delete, archive, and unarchive functionality directly within the student profile view.
    - [UPDATE] Improved Visibility for Archived Student Profile: The 'archived_at' column on the Student table will now appear whenever viewing of archived students.

commit 591ed4a761de5c64629e7fb5c949fcebf207e3f6
Author: Andrei Lazaro <alazaro.apps@gmail.com>
Date:   Tue Aug 15 03:19:33 2023 +0800

    Optimized the Student table and added an accessible Changelog link

    - [NEW] Changelog Accessibility: Added an accessible Changelog link
    - [NEW] Student Table Enhancement: Greater control over data with the addition of the "Show archived" toggle on the Student table.
    - [UPDATE]  Enhanced Student Table Filtering: Improved the student table with more precision and accurate filtering capabilities.

commit b089e86e77378da050241c14c091ff394d28a92c
Author: Andrei Lazaro <alazaro.apps@gmail.com>
Date:   Mon Aug 14 13:10:21 2023 +0800

     Embedded Sleekplan App Widget

    - [NEW] Embedded Sleekplan App Widget: The newly integrated Sleekplan App widget, will now effortlessly share user's thoughts, suggestions, and ideas directly within the app and stay up-to-date about the latest changes through it's intuitive and accessible widget.

commit 06ce715a79a84399136dd1ed91a6901a253867fe
Author: Andrei Lazaro <alazaro.apps@gmail.com>
Date:   Mon Aug 14 04:44:17 2023 +0800

    Updated Gradesheets table

    - [NEW] Show and hide gradesheets with no corresponding gradesheet pages.
    - [NEW] Added more gradesheets data filter capabilities.
    - Optimized Gradesheets table.
    - Bug fixes and various optimizations.

commit 5809733e2ae18c9b4936fcd29067f10d042effed
Author: Andrei Lazaro <alazaro.apps@gmail.com>
Date:   Mon Jul 31 18:10:34 2023 +0800

    Enhancement on Student table

commit e7618a706ee6a834f2a27885ab20fd2145e35718
Author: Andrei Lazaro <alazaro.apps@gmail.com>
Date:   Fri Jul 28 17:00:15 2023 +0800

    Bug fixes - Unable to delete grade sheets.

+++ app/Http/Controllers/Admin/UsersController.php
@@ -21,10 +21,10 @@ class UsersController extends Controller
 {
     public function index(Request $request)
     {
-        // abort_if(Gate::denies('user_access'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+        abort_if(Gate::denies('user_access'), Response::HTTP_FORBIDDEN, '403 Forbidden');
 
         if ($request->ajax()) {
-            $query = User::with(['roles'])->select(sprintf('%s.*', (new User)->table));
+            $query = User::with(['roles'])->approved();
 
             $table = Datatables::eloquent($query);
 
@@ -115,7 +115,7 @@ class UsersController extends Controller
 
     public function create()
     {
-        // abort_if(Gate::denies('user_create'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+        abort_if(Gate::denies('user_create'), Response::HTTP_FORBIDDEN, '403 Forbidden');
 
         $roles = Role::all()->pluck('title', 'id');
         $activityLogs = Activity::where('causer_type', 'App\Models\User')->orderBy('created_at', 'desc')->limit(4)->get();
@@ -143,7 +143,7 @@ class UsersController extends Controller
 
     public function edit(User $user)
     {
-        // abort_if(Gate::denies('user_edit'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+        abort_if(Gate::denies('user_edit'), Response::HTTP_FORBIDDEN, '403 Forbidden');
 
         $roles = Role::all()->pluck('title', 'id');
 
@@ -181,7 +181,7 @@ class UsersController extends Controller
 
     public function destroy(User $user)
     {
-        // abort_if(Gate::denies('user_delete'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+        abort_if(Gate::denies('user_delete'), Response::HTTP_FORBIDDEN, '403 Forbidden');
 
         $user->delete();
